+++ packages/components/src/components/tag-dismissible/tag-dismissible.tsx
@@ -1,4 +1,5 @@
 import { Component, Element, h, type JSX, Prop } from '@stencil/core';
+import type { PropTypes, SelectedAriaAttributes, Theme } from '../../types';
 import {
   AllowedTypes,
   attachComponentCss,
@@ -7,22 +8,14 @@ import {
   parseAndGetAriaAttributes,
   THEMES,
   validateProps,
-  warnIfDeprecatedPropValueIsUsed,
 } from '../../utils';
 import { getComponentCss } from './tag-dismissible-styles';
 import {
-  type TagDismissibleAriaAttribute,
-  type TagDismissibleColor,
-  type TagDismissibleColorDeprecated,
   TAG_DISMISSIBLE_ARIA_ATTRIBUTES,
   TAG_DISMISSIBLE_COLORS,
+  type TagDismissibleAriaAttribute,
+  type TagDismissibleColor,
 } from './tag-dismissible-utils';
-import type { PropTypes, SelectedAriaAttributes, Theme } from '../../types';
-
-type DeprecationMapType = Record<
-  TagDismissibleColorDeprecated,
-  Exclude<TagDismissibleColor, TagDismissibleColorDeprecated>
->;
 
 const propTypes: PropTypes<typeof TagDismissible> = {
   color: AllowedTypes.oneOf<TagDismissibleColor>(TAG_DISMISSIBLE_COLORS),
@@ -59,24 +52,7 @@ export class TagDismissible {
 
   public render(): JSX.Element {
     validateProps(this, propTypes);
-    const deprecationMap: DeprecationMapType = {
-      'background-default': 'background-base',
-    };
-    warnIfDeprecatedPropValueIsUsed<typeof TagDismissible, TagDismissibleColorDeprecated, TagDismissibleColor>(
-      this,
-      'color',
-      deprecationMap
-    );
-    attachComponentCss(
-      this.host,
-      getComponentCss,
-      (deprecationMap[this.color as keyof DeprecationMapType] || this.color) as Exclude<
-        TagDismissibleColor,
-        TagDismissibleColorDeprecated
-      >,
-      !!this.label,
-      this.theme
-    );
+    attachComponentCss(this.host, getComponentCss, this.color, !!this.label, this.theme);
 
     const PrefixedTagNames = getPrefixedTagNames(this.host);
     return (