+++ packages/components/src/components/tag/tag.spec.ts
@@ -1,6 +1,5 @@
 import * as getDirectChildHTMLElementUtils from '../../utils/dom/getDirectChildHTMLElement';
 import { Tag } from './tag';
-import * as warnIfDeprecatedPropValueIsUsed from '../../utils/log/warnIfDeprecatedPropValueIsUsed';
 
 describe('render', () => {
   it('should call getDirectChildHTMLElement() with correct parameters', () => {
@@ -12,23 +11,5 @@ describe('render', () => {
     component.render();
 
     expect(spy).toHaveBeenCalledWith(component.host, 'a,button');
-  });
-
-  it('should call warnIfDeprecatedPropValueIsUsed() with correct parameters', () => {
-    const spy = jest.spyOn(warnIfDeprecatedPropValueIsUsed, 'warnIfDeprecatedPropValueIsUsed');
-    const component = new Tag();
-    component.host = document.createElement('p-tag');
-    component.host.attachShadow({ mode: 'open' });
-
-    component.render();
-
-    expect(spy).toHaveBeenCalledWith(component, 'color', {
-      'background-default': 'background-base',
-      'neutral-contrast-high': 'primary',
-      'notification-neutral': 'notification-info-soft',
-      'notification-warning': 'notification-warning-soft',
-      'notification-success': 'notification-success-soft',
-      'notification-error': 'notification-error-soft',
-    });
   });
 });